shell command: chaning fixed_duration parameter data type to double (to support decimals)
Former-commit-id: 179d8997caf3a4dd658d52990a00e4c8a22d5a80

+++ fortscale/presidio-adapter/src/main/java/presidio/adapter/services/impl/AdapterExecutionServiceImpl.java
@@ -26,7 +26,7 @@ public class AdapterExecutionServiceImpl implements PresidioExecutionService {
     }
 
     @Override
-    public void run(DataSource dataSource, Instant startDate, Instant endDate, Long fixedDuration) throws Exception {
+    public void run(DataSource dataSource, Instant startDate, Instant endDate, Double fixedDuration) throws Exception {
         //todo: we need to consider doing the fetch & store at the same iteration
         logger.info("Start collector processing with params: data source:{}, from {}:{}, until {}:{}.",dataSource, CommonStrings.COMMAND_LINE_START_DATE_FIELD_NAME, startDate, CommonStrings.COMMAND_LINE_END_DATE_FIELD_NAME, endDate);
